While doing an update, try to handle disk space.  This means:
1) When downloading a file, raise an exception if doing so would
use up all the free space.
2) When applying an update, raise an exception if doing so would
use up all the space on the pool.  (This is only an estimate,
unfortunately.)
3) In both cases, if the filesystem is ZFS, and doing the operation
would raise the usage level above 85%, also consider this an error,
unless there is still a significant amount of space free (currently,
if the free space more than 4 times the required space).

Also add a "--force" option to freenas-update to force it to install
anyway.  The various methods/functions that do the downloading and
updating have a new defaults-to-false paremeter that allows this.

Ticket: #18020