When creating a BE for an update, turn off sync. Normally this could be dangerous, but we don't care about the BE until we acticate it, by which we will have turned sync back to the default.
On a FreeNAS Mini, this made an update run in about 9 minutes
instead of about 34 minutes.

I tried the same chagne with an ISO install, but did not get
any difference.  This may be due to amount of RAM, so I'll have
to try on a machine (or vm) with more memory.