Update Configuration.py
I was trying to installing TrueNAS core and was getting the "%{0} has invalid checksum" on the `base-os-(...).tgz`. It would be useful to be able to see what checksum it has calculated and what it's verifying against.

This could be moved into a debug log, but I've not found a way of enabling that during installation from a boot disk.